Changes for Version 2.0
These changes include everything done to make version 2.0 of the open
source license extraction tool. New readme and how to use files.
Explanation of the tagger package. The processing of files within file
directories is parallelized with go routines and channels. All notice
output to html is now alphabetized on the package level and on the
individual file level. The entire package has now gained the
capabilities to process packages downloaded with pip or gem. Added a
TODO file for future changes and other things that need done.